# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Todo API",
      "version": "1.0.0",
      "description": "Uma simples Todo API"
    },
    "paths": {
      "/tarefas": {
        "get": {
          "summary": "Lista todas as tarefas",
          "responses": {
            "200": {
              "description": "Lista todas as tarefas",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Tarefa"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Criar nova tarefa",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TarefaInput"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Tarefa criada",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Tarefa"
                  }
                }
              }
            }
          }
        }
      },
      "/tarefas/{id}": {
        "put": {
          "summary": "Atualizar tarefa",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TarefaInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Tarefa atualizada",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Tarefa"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "summary": "Deletar tarefa",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Tarefa deletada"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Tarefa": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "titulo": {
              "type": "string"
            },
            "completo": {
              "type": "boolean"
            }
          }
        },
        "TarefaInput": {
          "type": "object",
          "properties": {
            "titulo": {
              "type": "string"
            },
            "completo": {
              "type": "boolean"
            }
          }
        }
      }
    }
  }
  